+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Hook Groups.acceptMembershipMulti [#4268](https://github.com/rokwire/illinois-app/issues/4268).
+
 ## [6.0.31] - 2024-08-08
 ### Fixed
 - GroupsMembersPanel cleanup [#4268](https://github.com/rokwire/illinois-app/issues/4268).