+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Hook Groups.acceptMembershipMulti [#4268](https://github.com/rokwire/illinois-app/issues/4268).
 
 ## [6.1.6] - 2024-08-00
 ### Added
@@ -38,6 +40,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [6.1.0] - 2024-07-22
 ### Added
 - Created Public Surveys UI / round 0 [#4243](https://github.com/rokwire/illinois-app/issues/4243).
+
+## [6.0.32] - 2024-08-09
+### Added
+- Hook Groups.acceptMembershipMulti [#4268](https://github.com/rokwire/illinois-app/issues/4268).
 
 ## [6.0.31] - 2024-08-08
 ### Fixed